Generic blackboard implementation (#26)
Generalize blackboard. 

From user's perspective, in addition to being able to use generic
blackboard type, the user must now use tick directly with `bt.tick(...)`
instead of `bt.state.tick(...)` This change is required so that we can
pass a mutable reference of the blackboard to the user's tick callback
under the hood.