Return `None` if ticking `BT` after it finishes. (#45)
Today, ticking a behavior tree after it has finished has confusing
results. Looking at the code, it seems the intent was that once a
`State` finishes, it is not expected to tick again. For example, if you
tick a Sequence/Select after it has finished, it always returns
`Running`. Here's an example test that shows this strange behavior:

```rust
#[test]
fn weird_behavior() {
    let a = 4;

    let mut bt = BT::new(
        Select(vec![
            Invert(Box::new(Action(Dec))),
            Invert(Box::new(Action(Dec))),
            Invert(Box::new(Action(Dec))),
        ]),
        (),
    );

    let (a, s, _) = tick(a, 0.1, &mut bt);
    assert_eq!(a, 1);
    assert_eq!(s, Failure);
    let (a, s, _) = tick(a, 0.1, &mut bt);
    assert_eq!(a, 1);
    assert_eq!(s, Running);
    let (a, s, _) = tick(a, 0.1, &mut bt);
    assert_eq!(a, 1);
    assert_eq!(s, Running);
}
```

This is clearly wrong. We could fix this particular issue (make
Sequence/Select return a finished status if you tick it again), but it's
not clear that ticking a finished behavior makes sense at all - so
perhaps we should stop that.

This PR solves this by just keeping track of the BT returns a Success or
Failure status and then prevents ticking in those cases (returning an
Option)