Implement blackboard immutable reference function (#47)
First of all, thank you for creating such a great framework.

In the code I’m currently working on, I was using an older version of
bonsai-bt. However, as the version was updated, an issue arose.

The interface was modified to integrate the blackboard into the bt,
which caused problems with the existing logic that relied on an
immutable borrow of the blackboard. To address this, I added an
immutable reference function. Could you take a look and let me know if
it looks okay?